// src/app/modules/metadata/metadata.interface.ts
export interface IMetaData {
  activeUser?: number;
  blockedUser?: number;
  totalUser?: number;
  activeAdmin?: number;
  blockedAdmin?: number;
  totalAdmin?: number;
  upcomingTrip?: number;
  previousTrip?: number;
  activeTrip?: number;
  deletedTrip?: number;
  totalTrip?: number;
  tripApproved?: number;
  tripRejected?: number;
  tripPending?: number;
}